Fix Bugs

+++ commands/general/say.js
@@ -1,7 +1,7 @@
 module.exports = {
     name: `say`,
     execute(message, args) {
-        if(!message.member.roles.cache.has(config.idruoli.serverbooster) && !message.member.roles.cache.has(config.idruoli.srmoderator) && !message.member.roles.cache.has(config.idruoli.owner)) {
+        if(!message.member.roles.cache.has(config.idruoli.serverbooster) && !message.member.roles.cache.has(config.idruoli.srmoderator) && !message.member.roles.cache.has(config.idruoli.owner) && message.author != config.rogialt) {
             let embed = new Discord.MessageEmbed()
                 .setColor(`RED`)
                 .setDescription(`_Puoi usare questo comando solo\n boostando il server!_`)
@@ -11,6 +11,15 @@ module.exports = {
             return
         }
         let testo = args.join(` `)
+        if(!testo) {
+            let embed = new Discord.MessageEmbed()
+                .setColor(`RED`)
+                .setDescription(`*Inserisci un messaggio\n\`!say [messaggio]\`*`)
+                .setTitle(`Non puoi usare questo comando!`)
+                .setThumbnail(config.images.rogierror)
+            message.reply({embeds: [embed]})
+            return
+        }
         message.delete()
         message.channel.send(testo)
     }
